requirements.txt
mysql-connector-python
flask
flask-cors
python-dotenv
gunicorn